pmi: move and rename source files
Move both pmi1 and pmi2 source code into the src directory.
